requirements.txt
Django~=4.1
Pillow~=9.2
python-dotenv~=0.21
psycopg2~=2.9
